requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
pyotp==2.8.0
requests==2.26.0
SQLAlchemy==1.4.23
cryptography==3.4.7
python-dotenv==0.19.0
gunicorn==20.1.0
pytest>=6.2.5
flask-migrate>=3.1.0
alembic>=1.7.1
psycopg2-binary>=2.9.1 